Move CliOpts to client
