+++ engine/src/types.rs
@@ -4,12 +4,8 @@ use crate::network::rpc::Event;
 use crate::utils::{extract_app_lookup, extract_kate};
 use avail_core::DataLookup;
 use avail_subxt::{primitives::Header as DaHeader, utils::H256};
-use clap::{Parser, ValueEnum};
 use codec::{Decode, Encode};
-use color_eyre::{
-	eyre::{eyre, WrapErr},
-	Report, Result,
-};
+use color_eyre::{eyre::eyre, Report, Result};
 use kate_recovery::{
 	commitments,
 	matrix::{Dimensions, Partition},
@@ -21,7 +17,6 @@ use serde::{de::Error, Deserialize, Serialize};
 use sp_core::crypto::Ss58Codec;
 use sp_core::{blake2_256, bytes, ed25519};
 use std::fmt::{self, Display, Formatter};
-use std::fs;
 use std::num::{NonZeroU8, NonZeroUsize};
 use std::ops::Range;
 use std::str::FromStr;
@@ -45,60 +40,6 @@ pub const IDENTITY_PROTOCOL: &str = "/avail/light/1.0.0";
 pub const IDENTITY_AGENT_BASE: &str = "avail-light-client";
 pub const IDENTITY_AGENT_ROLE: &str = "light-client";
 pub const IDENTITY_AGENT_CLIENT_TYPE: &str = "rust-client";
-
-#[derive(Parser)]
-#[command(version)]
-pub struct CliOpts {
-	/// Path to the yaml configuration file
-	#[arg(short, long, value_name = "FILE")]
-	pub config: Option<String>,
-	/// Path to the toml identity file
-	#[arg(short, long, value_name = "FILE", default_value = "identity.toml")]
-	pub identity: String,
-	/// AppID for application client
-	#[arg(long, value_name = "app-id")]
-	pub app_id: Option<u32>,
-	/// Testnet or devnet selection
-	#[arg(short, long, value_name = "network")]
-	pub network: Option<Network>,
-	/// Run a clean light client, deleting existing avail_path folder
-	#[arg(long)]
-	pub clean: bool,
-	/// Enable finality sync
-	#[arg(short, long, value_name = "finality_sync_enable")]
-	pub finality_sync_enable: bool,
-	/// P2P port
-	#[arg(short, long)]
-	pub port: Option<u16>,
-	/// HTTP port
-	#[arg(long)]
-	pub http_server_port: Option<u16>,
-	/// Enable websocket transport
-	#[arg(long, value_name = "ws_transport_enable")]
-	pub ws_transport_enable: bool,
-	/// Log level
-	#[arg(long)]
-	pub verbosity: Option<LogLevel>,
-	// TODO: Deprecated since 1.9.0, remove it once it is safe
-	/// Avail secret seed phrase password, overrides password from identity file
-	#[arg(long)]
-	pub avail_passphrase: Option<String>,
-	/// Avail secret URI, overrides parameter from identity file
-	#[arg(long)]
-	pub avail_suri: Option<String>,
-	/// Seed string for libp2p keypair generation
-	#[arg(long)]
-	pub seed: Option<String>,
-	/// ed25519 private key for libp2p keypair generation
-	#[arg(long)]
-	pub private_key: Option<String>,
-	/// fraction and number of the block matrix part to fetch (e.g. 2/20 means second 1/20 part of a matrix) (default: None)
-	#[arg(long, value_parser = block_matrix_partition_format::parse)]
-	pub block_matrix_partition: Option<Partition>,
-	/// Set logs format to JSON
-	#[arg(long)]
-	pub logs_json: bool,
-}
 
 #[derive(Serialize, Deserialize, Debug)]
 #[serde(rename_all = "camelCase")]
@@ -663,12 +604,13 @@ impl From<&LibP2PConfig> for RocksDBStoreConfig {
 	}
 }
 
-impl From<&RuntimeConfig> for LibP2PConfig {
-	fn from(val: &RuntimeConfig) -> Self {
+impl From<(&RuntimeConfig, IdentifyConfig)> for LibP2PConfig {
+	fn from(pair: (&RuntimeConfig, IdentifyConfig)) -> Self {
+		let (val, identify) = pair;
 		Self {
 			secret_key: val.secret_key.clone(),
 			port: val.port,
-			identify: IdentifyConfig::new(),
+			identify,
 			autonat: val.into(),
 			kademlia: val.into(),
 			relays: val.relays.iter().map(Into::into).collect(),
@@ -791,11 +733,11 @@ impl FromStr for AgentVersion {
 }
 
 impl IdentifyConfig {
-	fn new() -> Self {
+	pub fn new(version: String) -> Self {
 		let agent_version = AgentVersion {
 			base_version: IDENTITY_AGENT_BASE.to_string(),
 			role: IDENTITY_AGENT_ROLE.to_string(),
-			release_version: clap::crate_version!().to_string(),
+			release_version: version,
 			client_type: IDENTITY_AGENT_CLIENT_TYPE.to_string(),
 		};
 
@@ -972,161 +914,11 @@ impl Default for RuntimeConfig {
 	}
 }
 
-#[derive(ValueEnum, Clone)]
-pub enum Network {
-	Local,
-	Hex,
-	Turing,
-}
-
-impl Network {
-	fn bootstrap_peer_id(&self) -> &str {
-		match self {
-			Network::Local => "12D3KooWStAKPADXqJ7cngPYXd2mSANpdgh1xQ34aouufHA2xShz",
-			Network::Hex => "12D3KooWBMwfo5qyoLQDRat86kFcGAiJ2yxKM63rXHMw2rDuNZMA",
-			Network::Turing => "12D3KooWBkLsNGaD3SpMaRWtAmWVuiZg1afdNSPbtJ8M8r9ArGRT",
-		}
-	}
-
-	fn bootstrap_multiaddrr(&self) -> &str {
-		match self {
-			Network::Local => "/ip4/127.0.0.1/tcp/39000",
-			Network::Hex => "/dns/bootnode.1.lightclient.hex.avail.so/tcp/37000",
-			Network::Turing => "/dns/bootnode.1.lightclient.turing.avail.so/tcp/37000",
-		}
-	}
-
-	fn full_node_ws(&self) -> &str {
-		match self {
-			Network::Local => "ws://127.0.0.1:9944",
-			Network::Hex => "wss://rpc-hex-devnet.avail.tools/ws",
-			Network::Turing => "wss://turing-rpc.avail.so/ws",
-		}
-	}
-
-	fn ot_collector_endpoint(&self) -> &str {
-		match self {
-			Network::Local => "http://127.0.0.1:4317",
-			Network::Hex => "http://otel.lightclient.hex.avail.so:4317",
-			Network::Turing => "http://otel.lightclient.turing.avail.so:4317",
-		}
-	}
-
-	fn genesis_hash(&self) -> &str {
-		match self {
-			Network::Local => "DEV",
-			Network::Hex => "9d5ea6a5d7631e13028b684a1a0078e3970caa78bd677eaecaf2160304f174fb",
-			Network::Turing => "d3d2f3a3495dc597434a99d7d449ebad6616db45e4e4f178f31cc6fa14378b70",
-		}
-	}
-
-	pub fn name(genesis_hash: &str) -> String {
-		let network = match genesis_hash {
-			"9d5ea6a5d7631e13028b684a1a0078e3970caa78bd677eaecaf2160304f174fb" => {
-				Network::Hex.to_string()
-			},
-			"d3d2f3a3495dc597434a99d7d449ebad6616db45e4e4f178f31cc6fa14378b70" => {
-				Network::Turing.to_string()
-			},
-			"DEV" => Network::Local.to_string(),
-			_ => "other".to_string(),
-		};
-
-		let prefix = &genesis_hash[..std::cmp::min(6, genesis_hash.len())];
-		format!("{}:{}", network, prefix)
-	}
-}
-
-impl fmt::Display for Network {
-	fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-		match self {
-			Network::Local => write!(f, "local"),
-			Network::Hex => write!(f, "hex"),
-			Network::Turing => write!(f, "turing"),
-		}
-	}
-}
-
-#[derive(ValueEnum, Clone)]
-pub enum LogLevel {
-	Info,
-	Debug,
-	Trace,
-	Warn,
-	Error,
-}
-
-impl Display for LogLevel {
-	fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-		f.write_str(match self {
-			LogLevel::Info => "INFO",
-			LogLevel::Debug => "DEBUG",
-			LogLevel::Trace => "TRACE",
-			LogLevel::Warn => "WARN",
-			LogLevel::Error => "ERROR",
-		})
-	}
-}
-
 impl RuntimeConfig {
 	/// A range bounded inclusively below and exclusively above
 	pub fn sync_range(&self, end: u32) -> Range<u32> {
 		let start = self.sync_start_block.unwrap_or(end);
 		Range { start, end }
-	}
-
-	pub fn load_runtime_config(&mut self, opts: &CliOpts) -> Result<()> {
-		if let Some(config_path) = &opts.config {
-			fs::metadata(config_path).map_err(|_| eyre!("Provided config file doesn't exist."))?;
-			let cfg: RuntimeConfig = confy::load_path(config_path)
-				.wrap_err(format!("Failed to load configuration from {}", config_path))?;
-			*self = cfg;
-		}
-
-		// Flags override the config parameters
-		if let Some(network) = &opts.network {
-			let bootstrap: (PeerId, Multiaddr) = (
-				PeerId::from_str(network.bootstrap_peer_id())
-					.wrap_err("unable to parse default bootstrap peerID")?,
-				Multiaddr::from_str(network.bootstrap_multiaddrr())
-					.wrap_err("unable to parse default bootstrap multi-address")?,
-			);
-			self.full_node_ws = vec![network.full_node_ws().to_string()];
-			self.bootstraps = vec![MultiaddrConfig::PeerIdAndMultiaddr(bootstrap)];
-			self.ot_collector_endpoint = network.ot_collector_endpoint().to_string();
-			self.genesis_hash = network.genesis_hash().to_string();
-		}
-
-		if let Some(loglvl) = &opts.verbosity {
-			self.log_level = loglvl.to_string();
-		}
-
-		if let Some(port) = opts.port {
-			self.port = port;
-		}
-		if let Some(http_port) = opts.http_server_port {
-			self.http_server_port = http_port;
-		}
-		self.sync_finality_enable |= opts.finality_sync_enable;
-		self.app_id = opts.app_id.or(self.app_id);
-		self.ws_transport_enable |= opts.ws_transport_enable;
-		if let Some(secret_key) = &opts.private_key {
-			self.secret_key = Some(SecretKey::Key {
-				key: secret_key.to_string(),
-			});
-		}
-
-		if let Some(seed) = &opts.seed {
-			self.secret_key = Some(SecretKey::Seed {
-				seed: seed.to_string(),
-			})
-		}
-
-		if let Some(partition) = &opts.block_matrix_partition {
-			self.block_matrix_partition = Some(*partition)
-		}
-
-		Ok(())
 	}
 }
